Crate pallet_spaces

Spaces Module

Spaces are the primary components of Subsocial. This module allows you to create a Space and customize it by updating its’ owner(s), content, unique handle, and permissions.

To understand how Spaces fit into the Subsocial ecosystem, you can think of how folders and files work in a file system. Spaces are similar to folders, that can contain Posts, in this sense. The permissions of the Space and Posts can be customized so that a Space could be as simple as a personal blog (think of a page on Facebook) or as complex as community (think of a subreddit) governed DAO.

Spaces can be compared to existing entities on web 2.0 platforms such as:

Blogs on Blogger,

Publications on Medium,

Groups or pages on Facebook,

Accounts on Twitter and Instagram,

Channels on YouTube,

Servers on Discord,

Forums on Discourse.

Modules

rpc

weights
THIS FILE WAS AUTO-GENERATED USING THE SUBSTRATE BENCHMARK CLI VERSION 2.0.1

Structs

GenesisConfig
Genesis config for the module, allow to build genesis storage.

Module

NextSpaceId
The next space id.

Space
Information about a space’s owner, its’ content, visibility and custom permissions.

SpaceById
Get the details of a space by its’ id.

SpaceIdByHandle
Find a given space id by its’ unique handle. If a handle is not registered, nothing will be returned (None).

SpaceIdsByOwner
Find the ids of all spaces owned, by a given account.

SpaceUpdate

Enums

Call
Dispatchable calls.

Error

RawEvent
Events for this module.

Constants

FIRST_SPACE_ID

RESERVED_SPACE_COUNT

Traits

AfterSpaceUpdated

BeforeSpaceCreated

Trait
The pallet’s configuration trait.

WeightInfo

Type Definitions

Event
RawEvent specialized for the configuration Trait